requirement.txt
pandas                      2.2.1
openpyxl                    3.1.2
regex                       2023.10.3
python-docx                 1.1.0